Rename a lot of things to hopefully be clearer
