+++ packages/malloy/src/lang/ast/field-space/query-spaces.ts
@@ -28,7 +28,6 @@ import {MalloyElement} from '../types/malloy-element';
 import {SpaceField} from '../types/space-field';
 
 import {WildcardFieldReference} from '../query-items/field-references';
-import {WildSpaceField} from './wild-space-field';
 import {RefinedSpace} from './refined-space';
 import {LookupResult} from '../types/lookup-result';
 import {ColumnSpaceField} from './column-space-field';
@@ -167,21 +166,17 @@ export abstract class QuerySpace
         }
         const fieldQueryDef = field.getQueryFieldDef(this.exprSpace);
         if (fieldQueryDef) {
-          if (field instanceof WildSpaceField) {
+          const typeDesc = field.typeDesc();
+          // Filter out fields whose type is 'error', which means that a totally bad field
+          // isn't sent to the compiler, where it will wig out.
+          // TODO Figure out how to make errors generated by `canContain` go in the right place,
+          // maybe by adding a logable element to SpaceFields.
+          if (
+            typeDesc.dataType !== 'error' &&
+            this.canContain(typeDesc) &&
+            !isEmptyNest(fieldQueryDef)
+          ) {
             fields.push(fieldQueryDef);
-          } else {
-            const typeDesc = field.typeDesc();
-            // Filter out fields whose type is 'error', which means that a totally bad field
-            // isn't sent to the compiler, where it will wig out.
-            // TODO Figure out how to make errors generated by `canContain` go in the right place,
-            // maybe by adding a logable element to SpaceFields.
-            if (
-              typeDesc.dataType !== 'error' &&
-              this.canContain(typeDesc) &&
-              !isEmptyNest(fieldQueryDef)
-            ) {
-              fields.push(fieldQueryDef);
-            }
           }
         }
         // TODO I removed the error here because during calculation of the refinement space,
