+++ packages/malloy/src/lang/ast/index.ts
@@ -22,12 +22,17 @@
  */
 
 export * from './query-properties/nest';
-export * from './elements/define-source';
-export * from './elements/define-query';
-export * from './elements/source';
-export * from './elements/refined-source';
-export * from './elements/source-query';
-export * from './elements/source-query-nodes';
+export * from './statements/define-source';
+export * from './statements/define-query';
+export * from './source-elements/source';
+export * from './source-elements/refined-source';
+export * from './source-query-elements/source-query-element';
+export * from './source-query-elements/sq-arrow';
+export * from './source-query-elements/sq-refine';
+export * from './source-query-elements/sq-source';
+export * from './source-query-elements/sq-reference';
+export * from './source-query-elements/sq-extend';
+
 export * from './source-properties/field-list-edit';
 export * from './source-properties/primary-key';
 export * from './source-properties/renames';
@@ -74,26 +79,29 @@ export * from './expressions/time-frame';
 export * from './expressions/top-by';
 export * from './expressions/unary';
 export * from './expressions/utils';
-export * from './elements/import-statement';
+export * from './statements/import-statement';
 export * from './query-properties/extend';
 export * from './parameters/has-parameter';
 export * from './query-elements/anonymous-query';
-export * from './query-elements/refine';
-export * from './query-elements/arrow';
-export * from './query-elements/view';
+export * from './query-elements/query-refine';
+export * from './query-elements/query-arrow';
+export * from './view-elements/view';
+export * from './view-elements/view-arrow';
+export * from './view-elements/view-refine';
 export * from './query-items/field-declaration';
 export * from './query-items/field-references';
 export * from './query-properties/aggregate';
 export * from './query-properties/calculate';
 export * from './query-properties/declare-fields';
-export * from './query-properties/dimensions';
+export * from './source-properties/dimensions';
 export * from './query-properties/extend';
 export * from './query-properties/filters';
 export * from './query-properties/group-by';
 export * from './query-properties/indexing';
-export * from './query-properties/joins';
+export * from './source-properties/joins';
+export * from './source-properties/view';
 export * from './query-properties/limit';
-export * from './query-properties/measures';
+export * from './source-properties/measures';
 export * from './query-properties/nest';
 export * from './query-properties/nests';
 export * from './query-properties/ordering';
@@ -101,11 +109,11 @@ export * from './query-properties/project-statement';
 export * from './query-properties/qop-desc';
 export * from './query-properties/sampling';
 export * from './query-properties/top';
-export * from './sources/named-source';
-export * from './sources/query-source';
-export * from './sources/from-sql-source';
-export * from './sources/sql-source';
-export * from './sources/table-source';
+export * from './source-elements/named-source';
+export * from './source-elements/query-source';
+export * from './source-elements/from-sql-source';
+export * from './source-elements/sql-source';
+export * from './source-elements/table-source';
 export * from './sql-elements/sql-statement';
 export * from './sql-elements/sql-string';
 export * from './types/annotation-elements';
